+++ packages/router/src/lib/operations/cancel.ts
@@ -1,6 +1,6 @@
 import {
   ajv,
-  getUuid,
+  createLoggingContext,
   InvariantTransactionData,
   InvariantTransactionDataSchema,
   RequestContext,
@@ -14,26 +14,21 @@ import { CancelInput, CancelInputSchema } from "../entities";
 export const cancel = async (
   invariantData: InvariantTransactionData,
   input: CancelInput,
-  requestContext: RequestContext,
+  _requestContext: RequestContext<string>,
 ): Promise<providers.TransactionReceipt | undefined> => {
-  const method = "cancel";
-  const methodId = getUuid();
+  const { requestContext, methodContext } = createLoggingContext(cancel.name, _requestContext);
 
   const { logger, contractWriter } = getContext();
-  logger.info({ method, methodId, requestContext, invariantData, input }, "Method start");
+  logger.info("Method start", requestContext, methodContext, { invariantData, input });
 
   // Validate InvariantData schema
   const validateInvariantData = ajv.compile(InvariantTransactionDataSchema);
   const validInvariantData = validateInvariantData(invariantData);
   if (!validInvariantData) {
     const error = validateInvariantData.errors?.map((err: any) => `${err.instancePath} - ${err.message}`).join(",");
-    logger.error(
-      { method, methodId, error: validateInvariantData.errors, invariantData },
-      "Invalid invariantData params",
-    );
     throw new ParamsInvalid({
-      method,
-      methodId,
+      methodContext,
+      invariantData,
       paramsError: error,
       requestContext,
     });
@@ -44,12 +39,11 @@ export const cancel = async (
   const validInput = validateInput(input);
   if (!validInput) {
     const error = validateInput.errors?.map((err: any) => `${err.instancePath} - ${err.message}`).join(",");
-    logger.error({ method, methodId, error: validateInput.errors, input }, "Invalid input params");
     throw new ParamsInvalid({
-      method,
-      methodId,
+      input,
       paramsError: error,
       requestContext,
+      methodContext,
     });
   }
 
@@ -63,10 +57,7 @@ export const cancel = async (
   }
 
   // Send to tx service
-  logger.info(
-    { method, methodId, requestContext, transactionId: invariantData.transactionId, side },
-    "Sending cancel tx",
-  );
+  logger.info("Sending cancel tx", requestContext, methodContext, { side });
 
   const receipt = await contractWriter.cancel(
     cancelChain,
@@ -76,6 +67,6 @@ export const cancel = async (
     },
     requestContext,
   );
-  logger.info({ method, methodId, requestContext, transactionHash: receipt.transactionHash }, "Method complete");
+  logger.info("Method complete", requestContext, methodContext, { transactionHash: receipt.transactionHash });
   return receipt;
 };