+++ packages/txservice/src/dispatch/dispatch.ts
@@ -1,7 +1,6 @@
 import { BigNumber, Signer } from "ethers";
-import { BaseLogger } from "pino";
 import PriorityQueue from "p-queue";
-import { delay, jsonifyError, mkAddress } from "@connext/nxtp-utils";
+import { createLoggingContext, delay, jsonifyError, Logger, mkAddress, RequestContext } from "@connext/nxtp-utils";
 
 import { Gas, WriteTransaction } from "../types";
 import { AlreadyMined, TransactionReplaced, TransactionReverted } from "../error";
@@ -39,7 +38,7 @@ export class TransactionDispatch extends ChainRpcProvider {
    * Centralized transaction monitoring class. Extends ChainRpcProvider, thus exposing all provider methods
    * through this class.
    *
-   * @param logger pino.BaseLogger used for logging.
+   * @param logger Logger used for logging.
    * @param signer Signer instance or private key used for signing transactions.
    * @param chainId The ID of the chain for which this class's providers will be servicing.
    * @param chainConfig Configuration for this specified chain, including the providers we'll
@@ -50,7 +49,7 @@ export class TransactionDispatch extends ChainRpcProvider {
    * configuration.
    */
   constructor(
-    logger: BaseLogger,
+    logger: Logger,
     signer: string | Signer,
     public readonly chainId: number,
     chainConfig: ChainConfig,
@@ -84,20 +83,20 @@ export class TransactionDispatch extends ChainRpcProvider {
    *
    * @returns Transaction instance with populated params, ready for submit.
    */
-  public async createTransaction(minTx: WriteTransaction): Promise<Transaction> {
+  public async createTransaction(minTx: WriteTransaction, context: RequestContext): Promise<Transaction> {
     // Make sure we haven't aborted dispatch.
     this.assertNotAborted();
     // Estimate gas here will throw if the transaction is going to revert on-chain for "legit" reasons. This means
     // that, if we get past this method, we can safely assume that the transaction will go through on submit, saving for
     // instances where the provider malfunctions.
-    const gas = await this.getGas(minTx);
+    const gas = await this.getGas(minTx, context);
     // Queue up the transaction with these values.
     const result = await this.queue.add(async (): Promise<{ value: Transaction | Error; success: boolean }> => {
       try {
         // NOTE: This call must be here, serialized within the queue, as it is dependent on current pending transaction count.
         const nonce = await this.getNonce();
         // Create a new transaction instance to track lifecycle. We will NOT be submitting here.
-        const transaction = new Transaction(this.logger, this, minTx, nonce, gas);
+        const transaction = new Transaction(this.logger, this, minTx, nonce, gas, undefined, context);
         this.buffer.insert(nonce, transaction);
         this.incrementNonce();
         return { value: transaction, success: true };
@@ -146,14 +145,14 @@ export class TransactionDispatch extends ChainRpcProvider {
   /**
    *
    */
-  private async monitorLoop() {
+  private async monitorLoop(context?: RequestContext) {
     // TODO: Throttle this loop during lulls in traffic, speed up during high load??
     if (this.isActive) {
       return;
     }
     this.isActive = true;
     while (this.shouldMonitor) {
-      await this.monitor();
+      await this.monitor(context);
       await delay(TransactionDispatch.MONITOR_POLL_PARITY);
     }
     this.isActive = false;
@@ -166,15 +165,16 @@ export class TransactionDispatch extends ChainRpcProvider {
    * Left exposed as public as it is useful for testing, and we may want to eventually execute
    * this on a whim (i.e. when a specific event occurs).
    */
-  public async monitor(): Promise<void> {
+  public async monitor(context?: RequestContext): Promise<void> {
     // Lazy solution: we only care about a potential hold-up if it could hold anything up.
+    const { requestContext, methodContext } = createLoggingContext(this.monitor.name, context);
     if (this.buffer.pending().length < 2) {
       await delay(TransactionDispatch.MONITOR_POLL_PARITY);
       return;
     }
     const result = await this.getTransactionCount();
     if (result.isErr()) {
-      this.logger.error({ err: jsonifyError(result.error) }, "Failed to get transaction count");
+      this.logger.error(`Failed to get transaction count`, requestContext, methodContext, jsonifyError(result.error));
       // TODO: If we keep getting failures due to RPC issue, escape out?
       return;
     }
@@ -194,7 +194,7 @@ export class TransactionDispatch extends ChainRpcProvider {
       // TODO: We need to verify that this is indeed a valid nonce gap.
 
       // If it is, then we need to backfill the gap.
-      await this.backfill(indexedNonce, undefined, "NOT_FOUND");
+      await this.backfill(indexedNonce, undefined, "NOT_FOUND", context);
     } else {
       if (tx.didFinish || tx.error) {
         // IF the transaction did finish already, we can ignore this one.
@@ -203,14 +203,14 @@ export class TransactionDispatch extends ChainRpcProvider {
       // Check to make sure that the transaction is not expired.
       if (tx.expired) {
         await tx.kill();
-        await this.backfill(indexedNonce, tx, "EXPIRED");
+        await this.backfill(indexedNonce, tx, "EXPIRED", context);
       } else if (tx.attempt > TransactionDispatch.TOO_MANY_ATTEMPTS) {
         // This will mark a transaction for death, but it does get 1 hail mary; the transaction
         // can still attempt to confirm whatever's currently been submitted.
         // TODO: Alternatively, we could give this tx a hail mary by allowing it to submit at max gas BEFORE
         // we kill it... ensuring that there is indeed no hope of getting it through before we give up entirely.
         await tx.kill();
-        await this.backfill(indexedNonce, tx, "TAKING_TOO_LONG");
+        await this.backfill(indexedNonce, tx, "TAKING_TOO_LONG", context);
       }
     }
   }
@@ -221,26 +221,22 @@ export class TransactionDispatch extends ChainRpcProvider {
    * @param blockade - The transaction that's blocking. Used predominantly for logging.
    * @param reason - The reason for the backfill, a string value to be specify in logs.
    */
-  private async backfill(nonce: number, blockade: Transaction | undefined, reason: string) {
+  private async backfill(nonce: number, blockade: Transaction | undefined, reason: string, context?: RequestContext) {
     // Make sure that the blockade transaction didn't manage to confirm.
     if (blockade?.didFinish) {
       return;
     }
+    const { requestContext, methodContext } = createLoggingContext(this.backfill.name, context);
 
-    const method = this.backfill.name;
     let addedToBuffer = false;
     try {
-      this.logger.error(
-        {
-          method,
-          nonce,
-          id: blockade?.id,
-          timestamp: blockade?.timestamp,
-          blockade: blockade?.params,
-          hashes: blockade?.responses.map((r) => r.hash),
-        },
-        `Transaction requires backfill: ${reason}`,
-      );
+      this.logger.warn(`Transaction requires backfill: ${reason}`, requestContext, methodContext, {
+        nonce,
+        id: blockade?.id,
+        timestamp: blockade?.timestamp,
+        blockade: blockade?.params,
+        hashes: blockade?.responses.map((r) => r.hash),
+      });
       const addressResult = await this.getAddress();
       if (addressResult.isErr()) {
         throw addressResult.error;
@@ -252,11 +248,11 @@ export class TransactionDispatch extends ChainRpcProvider {
         data: "0x",
         to: addressResult.value,
       };
-      const gas = await this.getGas(minTx);
+      const gas = await this.getGas(minTx, requestContext);
       // Set gas to maximum.
       gas.setToMax();
       // Create transaction, and forcefully overwrite the stale one (blockade) in the buffer.
-      const transaction = new Transaction(this.logger, this, minTx, nonce, gas, true);
+      const transaction = new Transaction(this.logger, this, minTx, nonce, gas, true, requestContext);
       const response = await this.sendTransaction(transaction);
       if (response.isErr()) {
         throw response.error;
@@ -271,30 +267,27 @@ export class TransactionDispatch extends ChainRpcProvider {
       }
       this.buffer.insert(nonce, transaction, true);
       addedToBuffer = true;
-      this.logger.info(
-        {
-          method,
-          nonce,
-          blockadeId: blockade?.id,
-          backfillTx: {
-            id: transaction.id,
-            hash: response.value.hash,
-            gasPrice: transaction.params.gasPrice.toString(),
-            gasLimit: transaction.params.gasLimit.toString(),
-            value: transaction.params.value.toString(),
-            to: transaction.params.to.toString(),
-          },
+      this.logger.info("Backfill completed successfully", requestContext, methodContext, {
+        nonce,
+        blockadeId: blockade?.id,
+        backfillTx: {
+          id: transaction.id,
+          hash: response.value.hash,
+          gasPrice: transaction.params.gasPrice.toString(),
+          gasLimit: transaction.params.gasLimit.toString(),
+          value: transaction.params.value.toString(),
+          to: transaction.params.to.toString(),
         },
-        "Backfill completed successfully",
-      );
+      });
     } catch (error) {
       if (error.type === AlreadyMined.type) {
         // We can assume that the transaction was sent using the signer outside of dispatch,
         // and as a result, we didn't have it stored in the buffer.
-        this.logger.warn(
-          { method, nonce, backfilledTxId: blockade?.id, error },
-          "Backfill failed: Transaction already mined",
-        );
+        this.logger.warn("Backfill failed: Transaction already mined", requestContext, methodContext, {
+          nonce,
+          backfilledTxId: blockade?.id,
+          error,
+        });
         if (!addedToBuffer) {
           // Fill the gap with a fake transaction. This tx will be labeled with .isBackfill = true,
           // and get removed as soon as the buffer trims up to its nonce.
@@ -312,6 +305,7 @@ export class TransactionDispatch extends ChainRpcProvider {
               nonce,
               new Gas(BigNumber.from(1), BigNumber.from(24001)),
               true,
+              requestContext,
             ),
             // overwrite
             true,
@@ -320,29 +314,20 @@ export class TransactionDispatch extends ChainRpcProvider {
         return;
       } else if (error.type === TransactionReplaced.type) {
         // The backfill was replaced by the original transaction, so we can just ignore it.
-        this.logger.info(
-          {
-            method,
-            nonce,
-            originalTxId: blockade?.id,
-            error,
-            replacement: error.replacement?.hash,
-            receipt: error.receipt?.hash,
-          },
-          "Backfill failed: Transaction replaced",
-        );
+        this.logger.info("Backfill failed: Transaction replaced", requestContext, methodContext, {
+          nonce,
+          originalTxId: blockade?.id,
+          error,
+          replacement: error.replacement?.hash,
+          receipt: error.receipt?.hash,
+        });
         return;
       }
       // Backfill failed, we should shut the system down.
-      this.logger.error(
-        {
-          method,
-          nonce,
-          backfilledTxId: blockade?.id,
-          error,
-        },
-        `Backfill failed: ${error.type}`,
-      );
+      this.logger.error(`Backfill failed: ${error.type}`, requestContext, methodContext, jsonifyError(error), {
+        nonce,
+        backfilledTxId: blockade?.id,
+      });
       // Raise the abort flag.
       // TODO / DEBUG: Temporarily disabled for debugging.
       // this.aborted = error;
@@ -359,8 +344,8 @@ export class TransactionDispatch extends ChainRpcProvider {
    *
    * @throws TransactionReverted if gas estimate fails.
    */
-  private async getGas(transaction: WriteTransaction): Promise<Gas> {
-    const method = this.getGas.name;
+  private async getGas(transaction: WriteTransaction, context?: RequestContext): Promise<Gas> {
+    const { requestContext, methodContext } = createLoggingContext(this.getGas.name, context);
     // Get gas estimate.
     let gasLimit: BigNumber;
     let result = await this.estimateGas(transaction);
@@ -371,21 +356,17 @@ export class TransactionDispatch extends ChainRpcProvider {
         // be included in the error.
         throw result.error;
       }
-      this.logger.warn(
-        {
-          method,
-          transaction: transaction,
-          error: result.error,
-        },
-        "Estimate gas failed due to an unexpected error.",
-      );
+      this.logger.warn("Estimate gas failed due to an unexpected error.", requestContext, methodContext, {
+        transaction: transaction,
+        error: jsonifyError(result.error),
+      });
       throw result.error;
     } else {
       gasLimit = result.value;
     }
 
     // Get gas price and create tracker instance.
-    result = await this.getGasPrice();
+    result = await this.getGasPrice(requestContext);
     if (result.isErr()) {
       throw result.error;
     }