+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - [router] Cancel sender transaction if subgraph is out of sync.
 - [router] Add admin endpoint to cancel sender txs.
+- [utils,sdk,router,txservice] Add custom logger class.
 - [contracts] Arbitrum deployment.
 - [utils] Sanitize signatures